+++ bus-gitlab/src/main/java/org/aoju/bus/gitlab/EpicsApi.java
@@ -124,7 +124,7 @@ public class EpicsApi extends AbstractApi {
      * @return a list of matching epics of the requested group and its subgroups
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Epic> getEpics(Object groupIdOrPath, Integer authorId, String labels, EpicOrderBy orderBy,
+    public List<Epic> getEpics(Object groupIdOrPath, Long authorId, String labels, EpicOrderBy orderBy,
                                SortOrder sortOrder, String search) throws GitLabApiException {
         return (getEpics(groupIdOrPath, authorId, labels, orderBy, sortOrder, search, getDefaultPerPage()).all());
     }
@@ -146,7 +146,7 @@ public class EpicsApi extends AbstractApi {
      * @return a list of matching epics of the requested group and its subgroups in the specified range
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Epic> getEpics(Object groupIdOrPath, Integer authorId, String labels,
+    public List<Epic> getEpics(Object groupIdOrPath, Long authorId, String labels,
                                EpicOrderBy orderBy, SortOrder sortOrder, String search, int page, int perPage) throws GitLabApiException {
         GitLabApiForm formData = new GitLabApiForm(page, perPage)
                 .withParam("author_id", authorId)
@@ -175,7 +175,7 @@ public class EpicsApi extends AbstractApi {
      * @return the Pager of matching epics of the requested group and its subgroups
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<Epic> getEpics(Object groupIdOrPath, Integer authorId, String labels,
+    public Pager<Epic> getEpics(Object groupIdOrPath, Long authorId, String labels,
                                 EpicOrderBy orderBy, SortOrder sortOrder, String search, int itemsPerPage) throws GitLabApiException {
         GitLabApiForm formData = new GitLabApiForm()
                 .withParam("author_id", authorId)
@@ -201,7 +201,7 @@ public class EpicsApi extends AbstractApi {
      * @return a Stream of matching epics of the requested group and its subgroups
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<Epic> getEpicsStream(Object groupIdOrPath, Integer authorId, String labels, EpicOrderBy orderBy,
+    public Stream<Epic> getEpicsStream(Object groupIdOrPath, Long authorId, String labels, EpicOrderBy orderBy,
                                        SortOrder sortOrder, String search) throws GitLabApiException {
         return (getEpics(groupIdOrPath, authorId, labels, orderBy, sortOrder, search, getDefaultPerPage()).stream());
     }
@@ -216,7 +216,7 @@ public class EpicsApi extends AbstractApi {
      * @return an Epic instance for the specified Epic
      * @throws GitLabApiException if any exception occurs
      */
-    public Epic getEpic(Object groupIdOrPath, Integer epicIid) throws GitLabApiException {
+    public Epic getEpic(Object groupIdOrPath, Long epicIid) throws GitLabApiException {
         Response response = get(Response.Status.OK, null, "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid);
         return (response.readEntity(Epic.class));
     }
@@ -230,7 +230,7 @@ public class EpicsApi extends AbstractApi {
      * @param epicIid       the IID of the epic to get
      * @return an Optional instance with the specified Epic as a value
      */
-    public Optional<Epic> getOptionalEpic(Object groupIdOrPath, Integer epicIid) {
+    public Optional<Epic> getOptionalEpic(Object groupIdOrPath, Long epicIid) {
         try {
             return (Optional.ofNullable(getEpic(groupIdOrPath, epicIid)));
         } catch (GitLabApiException glae) {
@@ -310,7 +310,7 @@ public class EpicsApi extends AbstractApi {
      * @return an Epic instance containing info on the newly created epic
      * @throws GitLabApiException if any exception occurs
      */
-    public Epic updateEpic(Object groupIdOrPath, Integer epicIid, String title, String labels, String description,
+    public Epic updateEpic(Object groupIdOrPath, Long epicIid, String title, String labels, String description,
                            Date startDate, Date endDate) throws GitLabApiException {
         Form formData = new GitLabApiForm()
                 .withParam("title", title, true)
@@ -341,7 +341,7 @@ public class EpicsApi extends AbstractApi {
      * @return an Epic instance containing info on the updated epic
      * @throws GitLabApiException if any exception occurs
      */
-    public Epic updateEpic(Object groupIdOrPath, Integer epicIid, Epic epic) throws GitLabApiException {
+    public Epic updateEpic(Object groupIdOrPath, Long epicIid, Epic epic) throws GitLabApiException {
         Form formData = new GitLabApiForm()
                 .withParam("title", epic.getTitle(), true)
                 .withParam("labels", epic.getLabels())
@@ -362,7 +362,7 @@ public class EpicsApi extends AbstractApi {
      * @param epicIid       the IID of the epic to delete
      * @throws GitLabApiException if any exception occurs
      */
-    public void deleteEpic(Object groupIdOrPath, Integer epicIid) throws GitLabApiException {
+    public void deleteEpic(Object groupIdOrPath, Long epicIid) throws GitLabApiException {
         delete(Response.Status.NO_CONTENT, null, "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid);
     }
 
@@ -376,7 +376,7 @@ public class EpicsApi extends AbstractApi {
      * @return a list of all epic issues belonging to the specified epic
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Epic> getEpicIssues(Object groupIdOrPath, Integer epicIid) throws GitLabApiException {
+    public List<Epic> getEpicIssues(Object groupIdOrPath, Long epicIid) throws GitLabApiException {
         return (getEpicIssues(groupIdOrPath, epicIid, getDefaultPerPage()).all());
     }
 
@@ -393,7 +393,7 @@ public class EpicsApi extends AbstractApi {
      * @return a list of all epic issues belonging to the specified epic in the specified range
      * @throws GitLabApiException if any exception occurs
      */
-    public List<Epic> getEpicIssues(Object groupIdOrPath, Integer epicIid, int page, int perPage) throws GitLabApiException {
+    public List<Epic> getEpicIssues(Object groupIdOrPath, Long epicIid, int page, int perPage) throws GitLabApiException {
         Response response = get(Response.Status.OK, getPageQueryParams(page, perPage), "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "issues");
         return (response.readEntity(new GenericType<List<Epic>>() {
         }));
@@ -410,7 +410,7 @@ public class EpicsApi extends AbstractApi {
      * @return the Pager of all epic issues belonging to the specified epic
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<Epic> getEpicIssues(Object groupIdOrPath, Integer epicIid, int itemsPerPage) throws GitLabApiException {
+    public Pager<Epic> getEpicIssues(Object groupIdOrPath, Long epicIid, int itemsPerPage) throws GitLabApiException {
         return (new Pager<Epic>(this, Epic.class, itemsPerPage, null, "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "issues"));
     }
 
@@ -424,7 +424,7 @@ public class EpicsApi extends AbstractApi {
      * @return a Stream of all epic issues belonging to the specified epic
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<Epic> getEpicIssuesStream(Object groupIdOrPath, Integer epicIid) throws GitLabApiException {
+    public Stream<Epic> getEpicIssuesStream(Object groupIdOrPath, Long epicIid) throws GitLabApiException {
         return (getEpicIssues(groupIdOrPath, epicIid, getDefaultPerPage()).stream());
     }
 
@@ -440,7 +440,7 @@ public class EpicsApi extends AbstractApi {
      * @return an EpicIssue instance containing info on the newly assigned epic issue
      * @throws GitLabApiException if any exception occurs
      */
-    public EpicIssue assignIssue(Object groupIdOrPath, Integer epicIid, Integer issueIid) throws GitLabApiException {
+    public EpicIssue assignIssue(Object groupIdOrPath, Long epicIid, Long issueIid) throws GitLabApiException {
         Response response = post(Response.Status.CREATED, (Form) null,
                 "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "issues", issueIid);
         return (response.readEntity(EpicIssue.class));
@@ -457,7 +457,7 @@ public class EpicsApi extends AbstractApi {
      * @return an EpicIssue instance containing info on the removed issue
      * @throws GitLabApiException if any exception occurs
      */
-    public EpicIssue removeIssue(Object groupIdOrPath, Integer epicIid, Integer issueIid) throws GitLabApiException {
+    public EpicIssue removeIssue(Object groupIdOrPath, Long epicIid, Long issueIid) throws GitLabApiException {
         Response response = delete(Response.Status.OK, null,
                 "groups", getGroupIdOrPath(groupIdOrPath), "epics", epicIid, "issues", issueIid);
         return (response.readEntity(EpicIssue.class));
@@ -476,7 +476,7 @@ public class EpicsApi extends AbstractApi {
      * @return an EpicIssue instance containing info on the newly assigned epic issue
      * @throws GitLabApiException if any exception occurs
      */
-    public EpicIssue updateIssue(Object groupIdOrPath, Integer epicIid, Integer issueIid, Integer moveBeforeId, Integer moveAfterId) throws GitLabApiException {
+    public EpicIssue updateIssue(Object groupIdOrPath, Long epicIid, Long issueIid, Long moveBeforeId, Long moveAfterId) throws GitLabApiException {
         GitLabApiForm form = new GitLabApiForm()
                 .withParam("move_before_id", moveBeforeId)
                 .withParam("move_after_id", moveAfterId);
